pyproject.toml
[tool.poetry]
name = "data_slacklake"
version = "0.1.0"
description = ""
authors = ["arthuroliveira44 <arthurbr.oliveira@gmail.com>"]
readme = "README.md"
packages = [{include = "data_slacklake"}]

[tool.poetry.dependencies]
python = "3.12.*"
slack-bolt = "^1.21.0"
databricks-sql-connector = "^3.6.0"
openai = "^1.59.0"
pydantic = "^2.10.0"
boto3 = "^1.42.27"
langchain = "1.2.6"
langchain-community = "^0.4"
numpy = "<2.0.0"
databricks-langchain = "0.13.0"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pre-commit = "^4.1.0"
pylint = "^3.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-p no:langsmith"
testpaths = ["tests"]
